executor: unstable test TestFlushTables move to standalone sandbox
