+++ executor/kvtest/kv_test.go
@@ -51,13 +51,3 @@ func TestDropStatsFromKV(t *testing.T) {
 	tk.MustQuery("select hist_id from mysql.stats_top_n where table_id = " + tblID).Check(
 		testkit.Rows())
 }
-
-func TestFlushTables(t *testing.T) {
-	store, clean := testkit.CreateMockStore(t)
-	defer clean()
-	tk := testkit.NewTestKit(t, store)
-
-	tk.MustExec("FLUSH TABLES")
-	err := tk.ExecToErr("FLUSH TABLES WITH READ LOCK")
-	require.Error(t, err)
-}
